move Perf to common class
